requirements.txt
alembic==1.4.2
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
decorator==4.4.2
fake-useragent==0.1.11
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.4.2
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
tornado==6.0.4
urllib3==1.25.8
wrapt==1.11.2
